Basic user permissions

+++ trydjango19/posts/views.py
@@ -1,6 +1,6 @@
 from urllib.parse import quote_plus
 
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect, Http404
 from django.core.paginator import Paginator
 from django.shortcuts import render, get_object_or_404, redirect
 from django.contrib import messages
@@ -21,6 +21,9 @@ def post_list(request):
     return render(request, 'posts/post_list.html', context=context)
 
 def post_create(request):
+    if not request.user.is_staff or not request.user.superuser:
+        raise Http404
+
     form = PostForm(request.POST or None, request.FILES or None)
     if form.is_valid():
         instance = form.save(commit=False)
@@ -43,6 +46,9 @@ def post_detail(request, slug=None):
     return render(request, 'posts/post_detail.html', {'context': context})
 
 def post_update(request, id=None):
+    if not request.user.is_staff or not request.user.superuser:
+        raise Http404
+
     instance = get_object_or_404(Post, id=id)
     form = PostForm(request.POST or None, request.FILES or None, instance=instance)
     if form.is_valid():
@@ -59,6 +65,9 @@ def post_update(request, id=None):
     return render(request, 'posts/post_form.html', context)
 
 def post_delete(request, id=None):
+    if not request.user.is_staff or not request.user.superuser:
+        raise Http404
+    
     instance = get_object_or_404(Post, id=id)
     instance.delete()
     messages.success(request, 'delete post succesfully')
